update: "기록 버그 수정 및 레이아웃 수정 및 등산 기록 히스토리에서 멤버를 누르면 해당 멤버 프로필로 이동"
Former-commit-id: 26618cf1b377365604e07a7a71c93dec6f25700d

+++ FrontEnd/app/src/main/java/com/sansantek/sansanmulmul/data/model/HikingHistory.kt
@@ -5,7 +5,7 @@ data class HikingHistory(
   val recordDistance: Double,
   val recordElevation: Double,
   val recordEndTime: String,
-  val recordKcal: Double,
+  val recordKcal: Int,
   val recordStartTime: String,
   val recordSteps: Int
 )
